Replace chest GUI, and refactor all the things
- Chest GUI is now replaced by a LogBot doppelganger. This lets us do
cool stuff like update the database when the chest is closed, rather
than frantrically watch it and update when the GUI disappears like an
idiot.
- Added highlighting to chest slots so you can tell what's changed at a
glance
- Added tooltips to detail what was in the slot previously and what's
there now
-  Removed key toggle. Why would you ever want to stop logging?
- More extensibility stuff. Now you should be able to register a new
table with DBThread.registerCreateTable(), then extend the Record class
and implement the necessarty methods to create a custom record type. Add
your record to the toDBqueue and it should be added to the database!
- Added record retrieval event. Currently this only works for chests,
but it should be possible to retrieve all records if the correct methods
are implemented.
- Reworked RecordChest and RecordBlock to suit the new extensibility
changes
- Added WaffleCore as a dependency for world UUID data